pyproject.toml
[project]
name = "trend-summary-mailer"
version = "0.1.0"
description = "A system to fetch daily trends from Qiita and Zenn, summarize them with LLM, and send them via email."
authors = [{ name = "Your Name", email = "your.email@example.com" }]
dependencies = [
    "feedparser",
    "google-generativeai",
    "sendgrid",
    "python-dotenv",
]
requires-python = ">=3.9"
readme = "README.md"
license = { text = "MIT" }

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["src"]
